feat(profile): header and profile feedback
- Add followers and following segment
- Separate edit-profile from profile
- Add logout functionality
